// web/src/types/dictionary.ts
export interface Dictionary {
  id?: number
  name?: string
  type?: string
  status?: 0 | 1
  remark?: string
  createBy?: string
  createAt?: string
}

export interface DictionaryData {
  id?: number
  label?: string
  value?: string
  type?: string
  status?: 0 | 1
  remark?: string
  createBy?: string
  createAt?: string
}